split web and kafka
