+++ src/main/java/eu/dissco/annotationprocessingservice/exception/ForbiddenException.java
@@ -1,0 +1,9 @@
+package eu.dissco.annotationprocessingservice.exception;
+
+public class ForbiddenException extends Exception {
+
+  public ForbiddenException(String annotationId, String creatorId){
+    super("No annotation with id " + annotationId +" found for user " + creatorId);
+  }
+
+}
